shader: Fix block emission and prefer it on GLES
On GLES Mesa rejects arrays of arrays as TCS and TES shader inputs and outputs.

This is based on section 4.3 of the GLSL 3.20 ES standard where an error should be
produced for "Per-vertex-arrayed arrays of arrays" and "For non-per-vertex-arrayed
array of arrays", so apparently something like

  in vec4 [][N];

is not allowed and on GLES interface blocks have to be used instead. In addition the
code needed some fixups when interface blocks are used.

v2: Emit correct type for elements in blocks

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: <Gurchetan Singh gurchetansingh@chromium.org>