vrend: remove duplicate VREND_TRANSFER_* defines
These already have to be the same as VIRGL_TRANSFER_{TO,FROM}_HOST, as
defined by the protocol. There's no point in having a separate set of
internal names for them.

While we're at it, rewrite tan if as a switch for clarity. This allows
us to detect internal usage of invalid values.

Signed-off-by: Erik Faye-Lund <erik.faye-lund@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>