vrend,feature: Add a caps flag for transform feedback 3
The piglits and KHR tests that use TF3 (i.e. multiple output
streams) tend to hang, and since the GLES host doesn't support
this but we lie about the GLSL version it is better to signal
the guest correctly that this is not supported.

v2: Set flag only on caps v2 code path, otherwise we might read
    from uninitialized memory
v3: Increment feature check version to indicate that the TF3
    feature flag is valid.

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>