ci: fold unreliable tests handing alongside the mustpass ones
There is no reason to handle the test separately, at the high-level at
least.

If needed we can make the results stand out better in the [near] future.

Note: this removes the (now dead) --compare-previous. I've considered
keeping it, yet since it doesn't attribute the flip-flopping tests it's
better to remove it for now.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Reviewed-by: Elie Tournier <elie.tournier@collabora.com>