ci: kill off custom --test
Unused and fairly convoluted for little reason. In practise we want a
full run when changes are made - fixing one test can easily regress
another.

Furthermore, this only works with deqp and does not throw a error when
attempted with piglit.

If needed we can reinstate this, it should handle both deqp and piglit.

Signed-off-by: Emil Velikov <emil.velikov@collabora.com>
Signed-off-by: Elie Tournier <elie.tournier@collabora.com>