configure: Don't enable valgrind unless specifically asked for
Running "make check" is quite slow if valgrind is enabled, and most of the time
failures is the result of a leak in the GL driver. Since when building packages
testing is enabled, this will make the package build fail when valgrind is
installed. So instead of automagically enabling valgrind when it is installed,
make it a configure option that also may fail if valgrind is not installed, but
requested.

v2: Use some lowercase variable to report enabled valgrind (Gurchetan)

Signed-off-by: Gert Wollny <gert.wollny@collabora.com>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>