virgl: add virgl_resource
Add virgl_resource as the type of global cross-renderer resources.
A virgl_resource cannot be used in a renderer context directly.
After a virgl_resource is attached to a renderer context, a local
context object can be created from the virgl_resource.  This is
known as "import".  A renderer context always imports global
resources and deals with context objects exclusively.

It is also possible to create a virgl_resource from a context
object.  This is known as "export".  Whether import or export, the
global resource and the context object have independent lifetimes.

There is only GL renderer right now, and the context object for GL
renderer is vrend_resource.  Importing from virgl_resource or
exporting to virgl_resource is a matter of increasing the refcount
of vrend_resource.  It is fast but can be a bit confusing.

In the future, a virgl_resource can be from a different source
(e.g., VK renderer or an external dmabuf).  Importing such a
virgl_resource as a vrend_resource will entail more works.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>