vtest: add experimental multi-client support
Add --multi-clients to enable experimental multi-client support.

Note that multi-client should be enabled only when the clients know
and trust each other.  This is because resources are global among
the clients.  Two clients trying to create two resources with the
same id will result in an id conflict for example.

Proper multi-client support requires changes to the protocol.  The
main change will be for the server to generate globally unique
resource ids and enforce access controls.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Reviewed-By: Gert Wollny <gert.wollny@collabora.com>
Reviewed-By: Gurchetan Singh <gurchetansingh@chromium.org>