virgl: make virgl_renderer_resource_{attach,detach}_iov generic
Add and use virgl_resource_{attach,detach}_iov.

When the iov of a virgl_resource is changed, we should in theory
notify all contexts where the virgl_resource has been attached to.
But because we plan to move to immutable iov for other renderers,
only vrend contexts need to be notified.  And because vrend contexts
use vrend_resource and refcount, we only need to update
vrend_resource.

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Tested-by: Gurchetan Singh <gurchetansingh@chromium.org>
Reviewed-by: Gurchetan Singh <gurchetansingh@chromium.org>