vrend: fix copy/inline transfer with gbm allocation
Like in vrend_renderer_transfer_internal, virgl_gbm_transfer is a
special path that is generally preferred.  However, a copy transfer
can be a synchronized transfer.  We don't want to use GBM in that
case.

v2: add a comment on glFinish
v3: use GBM only when VIRGL_TEXTURE_NEED_SWIZZLE is set

Signed-off-by: Chia-I Wu <olvaffe@gmail.com>
Reviewed-by: David Riley <davidriley@chromium.org>
Tested-by: David Riley <davidriley@chromium.org>
Acked-by: Gurchetan Singh <gurchetansingh@chromium.org>